+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+1.1.0 - (November 21, 2018)
+-----------------
+### Changed
+* Updated row striping logic to be more consistent on a per-section basis
+* Removed unnecessary logic when `fill` prop is not provided
+* Added checkmark icons to row selection column
+
 1.0.0 - (October 16, 2018)
 -----------------
 Initial stable release